additional cucumber scenario refs: https://github.com/metasfresh/metasfresh/issues/14683
