+++ backend/de.metas.business/src/main/java/de/metas/bpartner/blockstatus/interceptor/C_BPartner.java
@@ -20,12 +20,12 @@
  * #L%
  */
 
-package de.metas.bpartner.blockstatus;
+package de.metas.bpartner.blockstatus.interceptor;
 
-import de.metas.adempiere.model.I_C_Order;
 import de.metas.bpartner.BPartnerId;
-import de.metas.document.engine.DocStatus;
+import de.metas.bpartner.blockstatus.BPartnerBlockStatusService;
 import de.metas.i18n.AdMessageKey;
+import de.metas.interfaces.I_C_BPartner;
 import lombok.NonNull;
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
@@ -33,31 +33,30 @@ import org.adempiere.exceptions.AdempiereException;
 import org.compiere.model.ModelValidator;
 import org.springframework.stereotype.Component;
 
-@Interceptor(I_C_Order.class)
+@Interceptor(I_C_BPartner.class)
 @Component
-public class C_Order
+public class C_BPartner
 {
-	private static final AdMessageKey MSG_ORDER_WITH_BLOCKED_PARTNER = AdMessageKey.of("CannotCompleteOrderWithBlockedPartner");
+	private static final AdMessageKey MSG_CANNOT_ACTIVATE_BLOCKED_PARTNER = AdMessageKey.of("CannotActivateBlockedPartner");
 
 	private final BPartnerBlockStatusService bPartnerBlockStatusService;
 
-	public C_Order(@NonNull final BPartnerBlockStatusService bPartnerBlockStatusService)
+	public C_BPartner(@NonNull final BPartnerBlockStatusService bPartnerBlockStatusService)
 	{
 		this.bPartnerBlockStatusService = bPartnerBlockStatusService;
 	}
 
-	@ModelChange(timings = { ModelValidator.TYPE_BEFORE_NEW, ModelValidator.TYPE_BEFORE_CHANGE },
-			ifColumnsChanged = I_C_Order.COLUMNNAME_DocStatus)
-	public void validateBPartnerBlockedStatus(final I_C_Order order)
+	@ModelChange(timings = ModelValidator.TYPE_BEFORE_CHANGE, ifColumnsChanged = I_C_BPartner.COLUMNNAME_IsActive)
+	public void validateBPartnerBlockedStatus(@NonNull final I_C_BPartner bpartner)
 	{
-		if (!DocStatus.ofCode(order.getDocStatus()).isCompleted())
+		if (!bpartner.isActive())
 		{
 			return;
 		}
 
-		if (bPartnerBlockStatusService.isBPartnerBlocked(BPartnerId.ofRepoId(order.getC_BPartner_ID())))
+		if (bPartnerBlockStatusService.isBPartnerBlocked(BPartnerId.ofRepoId(bpartner.getC_BPartner_ID())))
 		{
-			throw new AdempiereException(MSG_ORDER_WITH_BLOCKED_PARTNER)
+			throw new AdempiereException(MSG_CANNOT_ACTIVATE_BLOCKED_PARTNER)
 					.markAsUserValidationError();
 		}
 	}
