+++ backend/de.metas.business/src/main/java/de/metas/bpartner/blockstatus/interceptor/C_Invoice.java
@@ -20,11 +20,13 @@
  * #L%
  */
 
-package de.metas.bpartner.blockstatus;
+package de.metas.bpartner.blockstatus.interceptor;
 
+import de.metas.adempiere.model.I_C_Invoice;
 import de.metas.bpartner.BPartnerId;
+import de.metas.bpartner.blockstatus.BPartnerBlockStatusService;
+import de.metas.document.engine.DocStatus;
 import de.metas.i18n.AdMessageKey;
-import de.metas.interfaces.I_C_BPartner;
 import lombok.NonNull;
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
@@ -32,30 +34,31 @@ import org.adempiere.exceptions.AdempiereException;
 import org.compiere.model.ModelValidator;
 import org.springframework.stereotype.Component;
 
-@Interceptor(I_C_BPartner.class)
+@Interceptor(I_C_Invoice.class)
 @Component
-public class C_BPartner
+public class C_Invoice
 {
-	private static final AdMessageKey MSG_CANNOT_ACTIVATE_BLOCKED_PARTNER = AdMessageKey.of("CannotActivateBlockedPartner");
+	private static final AdMessageKey MSG_INVOICE_WITH_BLOCKED_PARTNER = AdMessageKey.of("CannotCompleteInvoiceWithBlockedPartner");
 
 	private final BPartnerBlockStatusService bPartnerBlockStatusService;
 
-	public C_BPartner(@NonNull final BPartnerBlockStatusService bPartnerBlockStatusService)
+	public C_Invoice(@NonNull final BPartnerBlockStatusService bPartnerBlockStatusService)
 	{
 		this.bPartnerBlockStatusService = bPartnerBlockStatusService;
 	}
 
-	@ModelChange(timings = ModelValidator.TYPE_BEFORE_CHANGE, ifColumnsChanged = I_C_BPartner.COLUMNNAME_IsActive)
-	public void validateBPartnerBlockedStatus(@NonNull final I_C_BPartner bpartner)
+	@ModelChange(timings = { ModelValidator.TYPE_BEFORE_NEW, ModelValidator.TYPE_BEFORE_CHANGE },
+			ifColumnsChanged = I_C_Invoice.COLUMNNAME_DocStatus)
+	public void validateBPartnerBlockedStatus(@NonNull final I_C_Invoice invoiceRecord)
 	{
-		if (!bpartner.isActive())
+		if (!DocStatus.ofCode(invoiceRecord.getDocStatus()).isCompleted())
 		{
 			return;
 		}
 
-		if (bPartnerBlockStatusService.isBPartnerBlocked(BPartnerId.ofRepoId(bpartner.getC_BPartner_ID())))
+		if (bPartnerBlockStatusService.isBPartnerBlocked(BPartnerId.ofRepoId(invoiceRecord.getC_BPartner_ID())))
 		{
-			throw new AdempiereException(MSG_CANNOT_ACTIVATE_BLOCKED_PARTNER)
+			throw new AdempiereException(MSG_INVOICE_WITH_BLOCKED_PARTNER)
 					.markAsUserValidationError();
 		}
 	}
